requirements.txt
coinbase-advanced-py
numpy>=1.24.0
nltk
scikit-learn
pandas>=2.0.0
git+https://github.com/TA-Lib/ta-lib-python.git
# in linux do this instead:
# $ python -m pip install TA-Lib
xgboost
imblearn
matplotlib>=3.7.0
statsmodels
yfinance
newsapi-python
schedule
websockets<14
openai
tenacity
tweepy
flask
timeout-decorator
customtkinter
Pillow>=10.0.0  # Required for icon generation
pyobjc-framework-Cocoa>=11.0; sys_platform == 'darwin'  # Required for macOS icon support